requirements.txt
sounddevice==0.4.6
numpy==1.23.5
openai-whisper==20230314
